+++ nakago_ws/src/lib.rs
@@ -1,0 +1,11 @@
+//! # nakago-ws: A Warp HTTP routes integration for Nakago
+#![forbid(unsafe_code)]
+
+/// WebSocket Connections
+pub mod connections;
+
+/// The WebSocket Controller
+pub mod controller;
+
+pub use connections::Connections;
+pub use controller::{Controller, Handler};